chore: bump dependencies via bump-deps.sh (Issue #3604)
