+++ test/scripts/BumpDepsNoPrivateRepoReference.ts
@@ -1,21 +1,25 @@
 /**
- * Issue #3458 — live automation, contributor docs and the bump-deps test
- * comment must not reference the private `stSoftwareAU` issue trackers.
+ * Issue #3458 — the dependency-bump automation, its contributor docs and its
+ * tests must stay self-contained for public readers.
  *
- * `bump-deps.sh` ships in every public clone and prints two of these
- * references to stderr at runtime, so a `VibeCoding#NNNN` / `GRQ#NNNN` slug
- * points a public reader (or a public user running the script) at an issue
- * they cannot open. The behaviour being described — the automated
- * dependency-bump worker reverting a bump when a gate fails — is expressible
- * at concept level without a slug.
+ * `bump-deps.sh` ships in every public clone and printed two private
+ * issue-tracker slugs to stderr at runtime, so a public user running the script
+ * was pointed at issues they cannot open. The behaviour being described — the
+ * automated dependency-bump worker reverting a bump when an audit gate fails —
+ * is fully expressible at concept level.
  *
- * The test reads the real committed files and fails if any reintroduces a
- * private-repo issue slug, org-qualified slug or link. The public
- * `stSoftwareAU/NEAT-AI` repository is deliberately not matched.
+ * This guard reads the real committed files, and runs the script's failure path
+ * so the runtime message is checked too. Detection lives in
+ * `test/_privateRepoRefs.ts` and is unit-tested by
+ * `test/docs/PrivateRepoRefScanner.ts` (Issue #3604).
  */
 
-import { assertEquals } from "@std/assert";
+import { assert, assertEquals } from "@std/assert";
 import { fromFileUrl } from "@std/path";
+import {
+  findPrivateRepoRefs,
+  scanForPrivateRepoRefs,
+} from "../_privateRepoRefs.ts";
 
 /** Files named in the #3458 private-repo-reference audit. */
 const AUDITED_FILES = [
@@ -24,80 +28,93 @@ const AUDITED_FILES = [
   "./BumpDepsScript.ts",
 ];
 
-/**
- * Match a private `stSoftwareAU` repository reference: an issue-tracker slug
- * (`VibeCoding#1613`, `GRQ#3472`) or an org-qualified slug / link
- * (`stSoftwareAU/VibeCoding`, `github.com/stSoftwareAU/GRQ-cluster/...`).
- *
- * Case-sensitive on the upper-case repo tokens so lower-case in-tree fixtures
- * are never false positives.
- */
-const PRIVATE_REPO_PATTERN =
-  /(?:GRQ|VibeCoding)#\d+|stSoftwareAU\/(?:GRQ-cluster|GRQ-logs|GRQ|VibeCoding)/;
-
-/**
- * Return every 1-indexed line number carrying a private `stSoftwareAU`
- * repository issue slug, org-qualified slug or link.
- *
- * @param source raw file source
- * @returns 1-indexed line numbers with a private-repo reference
- */
-export function findPrivateRepoReferences(source: string): number[] {
-  const hits: number[] = [];
-  const lines = source.split("\n");
-  for (let i = 0; i < lines.length; i++) {
-    if (PRIVATE_REPO_PATTERN.test(lines[i])) {
-      hits.push(i + 1);
-    }
-  }
-  return hits;
-}
-
 for (const rel of AUDITED_FILES) {
-  Deno.test(`${rel} has no private repo reference (#3458)`, async () => {
+  Deno.test(`${rel} has no private repo reference (#3458)`, () => {
     const path = fromFileUrl(new URL(rel, import.meta.url));
-    const source = await Deno.readTextFile(path);
-    const hits = findPrivateRepoReferences(source);
+    const offenders = scanForPrivateRepoRefs([path]);
     assertEquals(
-      hits,
+      offenders,
       [],
-      `${rel} references a private stSoftwareAU repository on line(s) ` +
-        `${hits.join(", ")}. Reword to concept level so the file stays ` +
-        `meaningful to public readers.`,
+      `${rel} references a private stSoftwareAU repository (${
+        offenders.join("")
+      }). Reword to concept level so the file stays usable by public readers.`,
     );
   });
 }
 
-Deno.test("findPrivateRepoReferences flags a VibeCoding# slug", () => {
-  const src = 'echo "Worker should revert per VibeCoding#1613." >&2\n';
-  assertEquals(findPrivateRepoReferences(src), [1]);
+Deno.test("bump-deps.sh revert guidance stays concept level (#3458)", async () => {
+  const path = fromFileUrl(new URL("../../bump-deps.sh", import.meta.url));
+  const source = await Deno.readTextFile(path);
+  const revertLines = source
+    .split("\n")
+    .filter((line) => line.includes("should revert"));
+  assertEquals(
+    revertLines.length,
+    2,
+    "expected both audit-gate failure paths to print revert guidance",
+  );
+  for (const line of revertLines) {
+    assert(
+      line.includes("Worker should revert this bump."),
+      `runtime revert guidance must be self-contained, got: ${line.trim()}`,
+    );
+  }
 });
 
-Deno.test("findPrivateRepoReferences flags an org-qualified private repo", () => {
-  const src = [
-    "Intro line.",
-    "see https://github.com/stSoftwareAU/GRQ-cluster/blob/main/x.json here.",
-  ].join("\n");
-  assertEquals(findPrivateRepoReferences(src), [2]);
-});
+/**
+ * Run `bump-deps.sh` with a stub `deno` that fails the WASM smoke gate, so the
+ * script takes its audit-gate failure path and prints the revert advice.
+ *
+ * @returns the script's stderr and exit code
+ */
+async function runFailingSmokeGate(): Promise<{
+  stderr: string;
+  code: number;
+}> {
+  const home = await Deno.makeTempDir();
+  try {
+    const binDir = `${home}/bin`;
+    await Deno.mkdir(binDir, { recursive: true });
+    const stub = `${binDir}/deno`;
+    // `deno eval` (neatCore.rev read) succeeds; `deno test` (smoke gate) fails.
+    await Deno.writeTextFile(
+      stub,
+      '#!/bin/sh\ncase "$1" in\n  test) exit 1 ;;\n  *) exit 0 ;;\nesac\n',
+    );
+    await Deno.chmod(stub, 0o755);
 
-Deno.test("findPrivateRepoReferences ignores the public NEAT-AI repo", () => {
-  const src =
-    "Filed at https://github.com/stSoftwareAU/NEAT-AI/issues/3458 (public).\n";
-  assertEquals(findPrivateRepoReferences(src), []);
-});
+    const command = new Deno.Command("bash", {
+      args: ["./bump-deps.sh", "--no-internal", "--no-external"],
+      stdout: "piped",
+      stderr: "piped",
+      cwd: Deno.cwd(),
+      env: { PATH: `${binDir}:/usr/bin:/bin`, HOME: home },
+    });
+    const output = await command.output();
+    return {
+      stderr: new TextDecoder().decode(output.stderr),
+      code: output.code,
+    };
+  } finally {
+    await Deno.remove(home, { recursive: true });
+  }
+}
 
-Deno.test("findPrivateRepoReferences ignores a bare in-repo issue number", () => {
-  const src = "# Audit gate (two-phase, Issue #2465): reverts on failure.\n";
-  assertEquals(findPrivateRepoReferences(src), []);
-});
-
-Deno.test("findPrivateRepoReferences returns empty for concept-level prose", () => {
-  const src =
-    "Either gate failing exits non-zero and the worker reverts the bump.\n";
-  assertEquals(findPrivateRepoReferences(src), []);
-});
-
-Deno.test("findPrivateRepoReferences returns empty for empty input", () => {
-  assertEquals(findPrivateRepoReferences(""), []);
+Deno.test({
+  name:
+    "bump-deps.sh smoke-gate failure advises a revert without a private-repo slug (#3458)",
+  fn: async () => {
+    const result = await runFailingSmokeGate();
+    assertEquals(result.code, 1, `expected exit 1; stderr=${result.stderr}`);
+    assert(
+      result.stderr.includes("Worker should revert this bump."),
+      `expected concept-level revert advice; stderr=${result.stderr}`,
+    );
+    assertEquals(
+      findPrivateRepoRefs(result.stderr),
+      [],
+      `runtime output must not name the private orchestration repo; ` +
+        `stderr=${result.stderr}`,
+    );
+  },
 });
